+++ packages/eslint-plugin-baseui/__tests__/deprecated-theme-api-test.js
@@ -21,228 +21,758 @@ RuleTester.setDefaultConfig({
 
 const tests = {
   valid: [
-    // Valid deprecated property used as object property in styled object
     {
       code: `
-        const Foo = styled("div", {
-          border: "solid 1px black",
+        // TODO(should-fail)
+        // style function - destructure elsewhere in function
+        styled('div', ({$theme}) => {
+          const {colors} = $theme;
+          return {
+            color: colors.foreground,
+          };
         });
       `,
     },
-
-    // Valid deprecated property as object property in styled function
     {
       code: `
-        const Foo = styled("div", ({$theme}) => ({
-          background: $theme.colors.white,
-        }));
+        // TODO(should-fail)
+        // style function - using deprecated property in a nested function
+        styled('div', props => {
+          const foo = () => props.$theme.colors.foreground;
+          return {
+            color: foo(),
+          };
+        });
       `,
     },
-
-    // Valid deprecated property used from local variable
     {
       code: `
-        const foreground = "black";
-        const Foo = styled("div", {
-          color: foreground,
+        // TODO(should-fail)
+        // overrides - object defined elsewhere
+        const fooOverrides = {
+          Root: {
+            style: ({$theme}) => ({ color: $theme.colors.foreground }),
+          },
+        };
+        export default () => <Foo overrides={fooOverrides} />;
+      `,
+    },
+    {
+      code: `
+        // TODO(should-fail)
+        // useStyletron - direct array access, who would do this though?
+        function Foo({ children }) {
+          const boo = useStyletron()
+          return (
+            <div className={css({ color: boo[0].colors.foreground })}>{children}</div>
+          )
+        }
+      `,
+    },
+    {
+      code: `
+        // TODO(should-fail)
+        // useStyletron - destructure elsewhere in function
+        function Foo({ children }) {
+          const [css, theme] = useStyletron()
+          const {colors} = theme;
+          return (
+            <div className={css({ color: colors.foreground })}>{children}</div>
+          )
+        }
+      `,
+    },
+    {
+      code: `
+        // TODO(should-fail)
+        // createTheme - not passing an object literal
+        import {createTheme} from "baseui"
+        const primitives = {};
+        const overrides = { foreground: "#000" };
+        const theme = createTheme(primitives, overrides);
+      `,
+    },
+    {
+      code: `
+        // style function - using deprecated property on another object
+        styled('div', props => {
+          const foo = something.foreground;
+          return {
+            color: props.$theme.colors.primary,
+          };
         });
+      `,
+    },
+    {
+      code: `
+        // style function - renaming a valid property to one of our deprecated properties
+        styled('div', ({$theme: {colors: {foo: foreground}}}) => {
+          return {
+            color: foreground,
+          };
+        });
+      `,
+    },
+    {
+      code: `
+        // createTheme - not imported from baseui
+        function createTheme() {}
+        const theme = createTheme({}, { foreground: "#000" });
+      `,
+    },
+    {
+      code: `
+        // createTheme - no overrides at all
+        import {createTheme} from "baseui"
+        const theme = createTheme();
       `,
     },
   ],
   invalid: [
-    // Invalid use of a deprecated theme property via styled
     {
       code: `
-        const Foo = styled("div", props => ({
-          color: props.$theme.colors.foreground
-        }));
-      `,
-      errors: [{messageId: MESSAGES.replaceThemeProperty.id}],
-      output: `
-        const Foo = styled("div", props => ({
-          color: props.$theme.colors.contentPrimary
-        }));
-      `,
-    },
-
-    // Invalid use of a deprecated theme property via useStyletron
-    {
-      code: `
-        function Foo() {
-          const [css, theme] = useStyletron();
-          return (
-            <div
-              className={css({
-                color: theme.colors.foreground
-              })}
-            >
-              Foo
-            </div>
-          );
-        }
-      `,
-      errors: [{messageId: MESSAGES.replaceThemeProperty.id}],
-      output: `
-        function Foo() {
-          const [css, theme] = useStyletron();
-          return (
-            <div
-              className={css({
-                color: theme.colors.contentPrimary
-              })}
-            >
-              Foo
-            </div>
-          );
-        }
-      `,
-    },
-
-    // Invalid use of a deprecated theme property that is partially destructured via styled
-    {
-      code: `
-        const Foo = styled("div", ({ $theme }) => {
+        // styled - no destructuring
+        styled('div', props => {
           return {
-            color: $theme.colors.shadowFocus
+            color: props.$theme.colors.foreground,
           };
         });
       `,
-      errors: [{messageId: MESSAGES.deprecateThemeProperty.id}],
-    },
-
-    // Invalid use of a deprecated theme property that is partially destructured via withStyle
-    {
-      code: `
-        const Foo = withStyle(Foo, ({ $theme }) => {
+      output: `
+        // styled - no destructuring
+        styled('div', props => {
           return {
-            backgroundColor: $theme.colors.shadowFocus
+            color: props.$theme.colors.contentPrimary,
           };
         });
       `,
-      errors: [{messageId: MESSAGES.deprecateThemeProperty.id}],
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
     },
-
-    // Invalid use of a deprecated theme property that is partially destructured via overrides
     {
       code: `
-        function Foo() {
-          return (
-            <Button
-              overrides={{
-                Root: {
-                  style: ({ $theme }) => {
-                    return {
-                      color: $theme.colors.shadowFocus
-                    };
+        // withStyle - no destructuring
+        withStyle(<Foo />, props => {
+          return {
+            color: props.$theme.colors.foreground,
+          };
+        });
+      `,
+      output: `
+        // withStyle - no destructuring
+        withStyle(<Foo />, props => {
+          return {
+            color: props.$theme.colors.contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // styled - $theme destructured
+        styled('div', ({ $theme }) => {
+          return {
+            color: $theme.colors.foreground,
+          };
+        });
+      `,
+      output: `
+        // styled - $theme destructured
+        styled('div', ({ $theme }) => {
+          return {
+            color: $theme.colors.contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // withStyle - $theme destructured
+        withStyle(<Foo />, ({ $theme }) => {
+          return {
+            color: $theme.colors.foreground,
+          };
+        });
+      `,
+      output: `
+        // withStyle - $theme destructured
+        withStyle(<Foo />, ({ $theme }) => {
+          return {
+            color: $theme.colors.contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // styled - nested destructuring for "concern"
+        styled('div', ({ $theme: {colors} }) => {
+          return {
+            color: colors.foreground,
+          };
+        });
+      `,
+      output: `
+        // styled - nested destructuring for "concern"
+        styled('div', ({ $theme: {colors} }) => {
+          return {
+            color: colors.contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // withStyle - nested destructuring for "concern"
+        withStyle(<Foo />, ({ $theme: {colors} }) => {
+          return {
+            color: colors.foreground,
+          };
+        });
+      `,
+      output: `
+        // withStyle - nested destructuring for "concern"
+        withStyle(<Foo />, ({ $theme: {colors} }) => {
+          return {
+            color: colors.contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // styled - nested destructuring of the deprecated theme property
+        styled('div', ({ $theme: {colors: {foreground}} }) => {
+          return {
+            color: foreground,
+          };
+        });
+      `,
+      output: `
+        // styled - nested destructuring of the deprecated theme property
+        styled('div', ({ $theme: {colors: {contentPrimary}} }) => {
+          return {
+            color: contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // withStyle - nested destructuring of the deprecated theme property
+        withStyle(<Foo />, ({ $theme: {colors: {foreground}} }) => {
+          return {
+            color: foreground,
+          };
+        });
+      `,
+      output: `
+        // withStyle - nested destructuring of the deprecated theme property
+        withStyle(<Foo />, ({ $theme: {colors: {contentPrimary}} }) => {
+          return {
+            color: contentPrimary,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // styled - renaming a nested & destructured & deprecated theme property
+        styled('div', ({$theme: {colors: {foreground: foo}}}) => {
+          return {
+            color: foo,
+          };
+        });
+      `,
+      output: `
+        // styled - renaming a nested & destructured & deprecated theme property
+        styled('div', ({$theme: {colors: {contentPrimary: foo}}}) => {
+          return {
+            color: foo,
+          };
+        });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // overrides - no destructuring
+        export default () => (
+          <Foo
+            overrides={{
+              Root: { style: props => ({ color: props.$theme.colors.foreground }) }
+            }}
+          />
+        )
+      `,
+      output: `
+        // overrides - no destructuring
+        export default () => (
+          <Foo
+            overrides={{
+              Root: { style: props => ({ color: props.$theme.colors.contentPrimary }) }
+            }}
+          />
+        )
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // overrides - $theme destructured
+        export default () => (
+          <Foo
+            overrides={{
+              Root: { style: ({$theme}) => ({ color: $theme.colors.foreground }) }
+            }}
+          />
+        )
+      `,
+      output: `
+        // overrides - $theme destructured
+        export default () => (
+          <Foo
+            overrides={{
+              Root: { style: ({$theme}) => ({ color: $theme.colors.contentPrimary }) }
+            }}
+          />
+        )
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // overrides - "concern" destructured
+        export default () => (
+          <Foo
+            overrides={{
+              Root: { style: ({$theme: {colors}}) => ({ color: colors.foreground }) }
+            }}
+          />
+        )
+      `,
+      output: `
+        // overrides - "concern" destructured
+        export default () => (
+          <Foo
+            overrides={{
+              Root: { style: ({$theme: {colors}}) => ({ color: colors.contentPrimary }) }
+            }}
+          />
+        )
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // overrides - deprecated property destructured
+        export default () => (
+          <Foo
+            overrides={{
+              Root: {
+                style: ({
+                  $theme: {
+                    colors: { foreground }
                   }
-                }
-              }}
-            >
-              Foo
-            </Button>
-          );
-        }
-      `,
-      errors: [{messageId: MESSAGES.deprecateThemeProperty.id}],
-    },
-
-    // Invalid use of a deprecated theme property that is partially destructured via styled
-    {
-      code: `
-        const Foo = styled("div", ({ $theme }) => {
-          const [colors] = $theme;
-          return {
-            color: colors.shadowFocus
-          };
-        });
-      `,
-      errors: [{messageId: 'deprecateThemeProperty'}],
-    },
-    {
-      code: `
-        const Foo = styled("div", ({$theme: {colors: {shadowFocus}}}) => ({
-          color: shadowFocus
-        }));
-      `,
-      errors: [
-        {messageId: MESSAGES.deprecateThemeProperty.id},
-        {messageId: MESSAGES.deprecateThemeProperty.id},
-      ],
-    },
-
-    // Invalid use of a deprecated theme property that is fully destructured via useStyletron
-    {
-      code: `
-        function Foo() {
-          const [
-            css,
-            {
-              colors: { shadowFocus }
-            }
-          ] = useStyletron();
-          return (
-            <div
-              className={css({
-                color: shadowFocus
-              })}
-            >
-              Foo
-            </div>
-          );
-        }
-      `,
-      errors: [
-        {messageId: MESSAGES.deprecateThemeProperty.id},
-        {messageId: MESSAGES.deprecateThemeProperty.id},
-      ],
-    },
-
-    // Invalid use of old typography values
-    {
-      code: `
-        const Foo = styled("div", ({ $theme }) => ({
-          ...$theme.typography.font300
-        }));
-      `,
-      errors: [{messageId: MESSAGES.replaceThemeProperty.id}],
-      output: `
-        const Foo = styled("div", ({ $theme }) => ({
-          ...$theme.typography.ParagraphMedium
-        }));
-      `,
-    },
-
-    // Invalid setting deprecated property in createTheme
-    {
-      code: `
-        import { createTheme } from "baseui";
-        const theme = createTheme(
-          {},
-          {
-            colors: {
-              shadowFocus: "pink"
-            }
-          }
-        );      
-      `,
-      errors: [{messageId: MESSAGES.deprecateThemeProperty.id}],
-    },
-
-    // Invalid within class component method
-    {
-      code: `
-        class Foo extends React.Component {
-          render() {
-            const Boo = styled('div', ({$theme}) => {
-              return {
-                backgroundColor: $theme.colors.background
+                }) => ({ color: foreground })
               }
-            });
-            return <Boo />;
-          }
+            }}
+          />
+        )
+      `,
+      output: `
+        // overrides - deprecated property destructured
+        export default () => (
+          <Foo
+            overrides={{
+              Root: {
+                style: ({
+                  $theme: {
+                    colors: { contentPrimary }
+                  }
+                }) => ({ color: contentPrimary })
+              }
+            }}
+          />
+        )
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // overrides - deprecated property destructured and renamed
+        export default () => (
+          <Foo
+            overrides={{
+              Root: {
+                style: ({
+                  $theme: {
+                    colors: { foreground: foo }
+                  }
+                }) => ({ color: foo })
+              }
+            }}
+          />
+        )
+      `,
+      output: `
+        // overrides - deprecated property destructured and renamed
+        export default () => (
+          <Foo
+            overrides={{
+              Root: {
+                style: ({
+                  $theme: {
+                    colors: { contentPrimary: foo }
+                  }
+                }) => ({ color: foo })
+              }
+            }}
+          />
+        )
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // useStyletron - basic
+        function Foo({ children }) {
+          const [css, theme] = useStyletron()
+          return (
+            <div className={css({ color: theme.colors.foreground })}>{children}</div>
+          )
         }
       `,
-      errors: [{messageId: MESSAGES.replaceThemeProperty.id}],
+      output: `
+        // useStyletron - basic
+        function Foo({ children }) {
+          const [css, theme] = useStyletron()
+          return (
+            <div className={css({ color: theme.colors.contentPrimary })}>{children}</div>
+          )
+        }
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // useStyletron - alternate name for theme
+        function Foo({ children }) {
+          const [css, foo] = useStyletron()
+          return (
+            <div className={css({ color: foo.colors.foreground })}>{children}</div>
+          )
+        }
+      `,
+      output: `
+        // useStyletron - alternate name for theme
+        function Foo({ children }) {
+          const [css, foo] = useStyletron()
+          return (
+            <div className={css({ color: foo.colors.contentPrimary })}>{children}</div>
+          )
+        }
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // useStyletron - destructure "concern"
+        function Foo({ children }) {
+          const [css, {colors}] = useStyletron()
+          return (
+            <div className={css({ color: colors.foreground })}>{children}</div>
+          )
+        }
+      `,
+      output: `
+        // useStyletron - destructure "concern"
+        function Foo({ children }) {
+          const [css, {colors}] = useStyletron()
+          return (
+            <div className={css({ color: colors.contentPrimary })}>{children}</div>
+          )
+        }
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // useStyletron - destructure "concern" and rename
+        function Foo({ children }) {
+          const [css, {colors: foo}] = useStyletron()
+          return (
+            <div className={css({ color: foo.foreground })}>{children}</div>
+          )
+        }
+      `,
+      output: `
+        // useStyletron - destructure "concern" and rename
+        function Foo({ children }) {
+          const [css, {colors: foo}] = useStyletron()
+          return (
+            <div className={css({ color: foo.contentPrimary })}>{children}</div>
+          )
+        }
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // useStyletron - destructure deprecated property
+        function Foo({ children }) {
+          const [css, {colors: {foreground}}] = useStyletron()
+          return (
+            <div className={css({ color: foreground })}>{children}</div>
+          )
+        }
+      `,
+      output: `
+        // useStyletron - destructure deprecated property
+        function Foo({ children }) {
+          const [css, {colors: {contentPrimary}}] = useStyletron()
+          return (
+            <div className={css({ color: contentPrimary })}>{children}</div>
+          )
+        }
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // useStyletron - destructure deprecated property with renaming
+        function Foo({ children }) {
+          const [css, {colors: {foreground: foo}}] = useStyletron()
+          return (
+            <div className={css({ color: foo })}>{children}</div>
+          )
+        }
+      `,
+      output: `
+        // useStyletron - destructure deprecated property with renaming
+        function Foo({ children }) {
+          const [css, {colors: {contentPrimary: foo}}] = useStyletron()
+          return (
+            <div className={css({ color: foo })}>{children}</div>
+          )
+        }
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+        // createTheme - setting a deprecated property
+        import {createTheme} from "baseui";
+        const theme = createTheme({}, { foreground: "#000" });
+      `,
+      output: `
+        // createTheme - setting a deprecated property
+        import {createTheme} from "baseui";
+        const theme = createTheme({}, { contentPrimary: "#000" });
+      `,
+      errors: [
+        {
+          messageId: MESSAGES.replaceThemeProperty.id,
+          data: {
+            old: 'foreground',
+            new: 'contentPrimary',
+          },
+        },
+      ],
     },
   ],
 };
