+++ packages/eslint-plugin-baseui/src/index.js
@@ -7,7 +7,6 @@ LICENSE file in the root directory of this source tree.
 
 'use strict';
 
-const Theme = require('./theme.js');
 const DeprecatedThemeAPI = require('./deprecated-theme-api.js');
 const DeprecatedComponentAPI = require('./deprecated-component-api.js');
 const NoDeepImports = require('./no-deep-imports.js');
@@ -17,6 +16,5 @@ module.exports = {
     'deprecated-theme-api': DeprecatedThemeAPI,
     'deprecated-component-api': DeprecatedComponentAPI,
     'no-deep-imports': NoDeepImports,
-    theme: Theme,
   },
 };
